move login logic to hook
