+++ src/components/auth/LoginButton.tsx
@@ -1,31 +1,16 @@
 'use client';
 
-import { useRouter } from 'next/navigation';
 import { Button, Text } from '@kuma-ui/core';
-import { SpotifyApi } from '@spotify/web-api-ts-sdk';
 
-import {
-  REROUTE_HOME_URL,
-  SPOTIFY_CLIENT_ID,
-  SPOTIFY_POSTBACK_URL,
-  SPOTIFY_SCOPES,
-} from '@/spotify/constants';
+import useLogin from '@/hooks/login';
 
 export interface LoginButtonProps {}
 
 export const LoginButton: React.FC<LoginButtonProps> = () => {
-  const router = useRouter();
+  const login = useLogin();
 
   return (
-    <Button
-      onClick={() =>
-        SpotifyApi.performUserAuthorization(
-          SPOTIFY_CLIENT_ID,
-          REROUTE_HOME_URL,
-          SPOTIFY_SCOPES,
-          SPOTIFY_POSTBACK_URL,
-        ).then(() => router.refresh())
-      }>
+    <Button onClick={() => login()}>
       <Text fontSize={24}>{'Login'}</Text>
     </Button>
   );
